# assets/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample promtpay QR generator in GO",
        "title": "Promptpay QR Services",
        "contact": {
            "name": "Kawin Viriyaprasopsook",
            "url": "https://github.com/zercle/promtpay-qr-services",
            "email": "kawin.vir@zercle.tech"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/qr/billpayment": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json",
                    "image/png"
                ],
                "parameters": [
                    {
                        "maximum": 99999999999.99,
                        "minimum": 0,
                        "type": "number",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "invoice",
                        "in": "query"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "merchan_name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "static",
                            "dynamic"
                        ],
                        "type": "string",
                        "name": "qr_type",
                        "in": "query"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "ref1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "ref2",
                        "in": "query"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "ref3",
                        "in": "query"
                    },
                    {
                        "maxLength": 2,
                        "type": "string",
                        "name": "suffix",
                        "in": "query"
                    },
                    {
                        "maxLength": 13,
                        "minLength": 10,
                        "type": "string",
                        "name": "tax_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "response by QR image",
                        "name": "image",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "image/png"
                ],
                "parameters": [
                    {
                        "description": "transaction detail",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QrRequest"
                        }
                    },
                    {
                        "type": "boolean",
                        "description": "response by QR image",
                        "name": "image",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/v1/qr/mypromptqr": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "maximum": 99999999999.99,
                        "minimum": 0,
                        "type": "number",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "invoice",
                        "in": "query"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "merchan_name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "static",
                            "dynamic"
                        ],
                        "type": "string",
                        "name": "qr_type",
                        "in": "query"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "ref1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "ref2",
                        "in": "query"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "ref3",
                        "in": "query"
                    },
                    {
                        "maxLength": 2,
                        "type": "string",
                        "name": "suffix",
                        "in": "query"
                    },
                    {
                        "maxLength": 13,
                        "minLength": 10,
                        "type": "string",
                        "name": "tax_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "response by QR image",
                        "name": "image",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "transaction detail",
                        "name": "post_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QrRequest"
                        }
                    },
                    {
                        "type": "boolean",
                        "description": "response by QR image",
                        "name": "image",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/v1/qr/qr30": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "maximum": 99999999999.99,
                        "minimum": 0,
                        "type": "number",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "invoice",
                        "in": "query"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "merchan_name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "static",
                            "dynamic"
                        ],
                        "type": "string",
                        "name": "qr_type",
                        "in": "query"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "ref1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "ref2",
                        "in": "query"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "ref3",
                        "in": "query"
                    },
                    {
                        "maxLength": 2,
                        "type": "string",
                        "name": "suffix",
                        "in": "query"
                    },
                    {
                        "maxLength": 13,
                        "minLength": 10,
                        "type": "string",
                        "name": "tax_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "response by QR image",
                        "name": "image",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "transaction detail",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QrRequest"
                        }
                    },
                    {
                        "type": "boolean",
                        "description": "response by QR image",
                        "name": "image",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/api/v1/qr/qrcs": {
            "get": {
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "maximum": 99999999999.99,
                        "minimum": 0,
                        "type": "number",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "invoice",
                        "in": "query"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "merchan_name",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "static",
                            "dynamic"
                        ],
                        "type": "string",
                        "name": "qr_type",
                        "in": "query"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "ref1",
                        "in": "query",
                        "required": true
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "ref2",
                        "in": "query"
                    },
                    {
                        "maxLength": 20,
                        "type": "string",
                        "name": "ref3",
                        "in": "query"
                    },
                    {
                        "maxLength": 2,
                        "type": "string",
                        "name": "suffix",
                        "in": "query"
                    },
                    {
                        "maxLength": 13,
                        "minLength": 10,
                        "type": "string",
                        "name": "tax_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "response by QR image",
                        "name": "image",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "transaction detail",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.QrRequest"
                        }
                    },
                    {
                        "type": "boolean",
                        "description": "response by QR image",
                        "name": "image",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "helpers.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "source": {},
                "title": {
                    "type": "string"
                }
            }
        },
        "helpers.ResponseForm": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/helpers.ResponseError"
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "result": {},
                "result_info": {
                    "$ref": "#/definitions/helpers.ResultInfo"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "helpers.ResultInfo": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "models.QrRequest": {
            "type": "object",
            "required": [
                "amount",
                "ref1",
                "tax_id"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "maximum": 99999999999.99,
                    "minimum": 0
                },
                "invoice": {
                    "type": "string",
                    "maxLength": 20
                },
                "merchan_name": {
                    "type": "string",
                    "maxLength": 20
                },
                "qr_type": {
                    "type": "string",
                    "enum": [
                        "static",
                        "dynamic"
                    ]
                },
                "ref1": {
                    "type": "string",
                    "maxLength": 20
                },
                "ref2": {
                    "type": "string",
                    "maxLength": 20
                },
                "ref3": {
                    "type": "string",
                    "maxLength": 20
                },
                "suffix": {
                    "type": "string",
                    "maxLength": 2
                },
                "tax_id": {
                    "type": "string",
                    "maxLength": 13,
                    "minLength": 10
                }
            }
        },
        "models.QrResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "data": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "invoice": {
                    "type": "string"
                },
                "merchan_name": {
                    "type": "string"
                },
                "qr_type": {
                    "type": "string",
                    "enum": [
                        "static",
                        "dynamic"
                    ]
                },
                "ref1": {
                    "type": "string"
                },
                "ref2": {
                    "type": "string"
                },
                "ref3": {
                    "type": "string"
                },
                "suffix": {
                    "type": "string"
                },
                "tax_id": {
                    "type": "string"
                }
            }
        }
    }
}